Timmy beta3.1.2
- Changed file structure for Ticket Systems
- Added some more cts commands.
- Added the feature to skip using forms and instantly open a ticket on button click (cts feature).  
- We now run on port 80. (Use sudo/admin since port 80 is a privileged port)
