+++ main.py
@@ -199,7 +199,7 @@ async def info():
 
 if __name__ == "__main__":
     if os.getenv("StartAPI"):
-        uvicorn.run(app, host="0.0.0.0", port=443)
+        uvicorn.run(app, host="0.0.0.0", port=80)
     else:
         bot.run(os.getenv("TOKEN"))
