+++ utils/bots/TicketSystem/ticket_control.py
@@ -1,6 +1,9 @@
+from __future__ import annotations
+
 import asyncio
 
 import discord
+from discord import app_commands
 from discord.ext import commands
 
 from core import database
