+++ utils/bots/TicketSystem/tickets/hr_system.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from typing import Literal
 
 import discord
-from discord import app_commands, ui
+from discord import app_commands
 from discord.ext import commands
 from googleapiclient.discovery import build
 from googleapiclient.errors import HttpError
@@ -40,198 +40,6 @@ orgUnit = {
 
 creds = access_secret("adm_t", True, 0, SCOPES)
 service = build("admin", "directory_v1", credentials=creds)
-
-
-class GSuiteForm(ui.Modal, title="GSuite Account/Email Request"):
-    def __init__(self, bot: commands.Bot) -> None:
-        super().__init__(timeout=None)
-        self.bot = bot
-
-    firstName = ui.TextInput(
-        label="What is your first name?",
-        style=discord.TextStyle.short,
-        max_length=100,
-    )
-
-    lastName = ui.TextInput(
-        label="What is your last name?",
-        style=discord.TextStyle.short,
-        max_length=1024,
-    )
-
-    descriptionTI = ui.TextInput(
-        label="Personal Email Address",
-        style=discord.TextStyle.short,
-        placeholder="This is not required but if given, you will get a copy of sign in details in your email.",
-        max_length=1024,
-        required=False,
-    )
-
-    dept = ui.TextInput(
-        label="Primary Department",
-        style=discord.TextStyle.short,
-        placeholder="What is the current department you are in?",
-        max_length=1024,
-    )
-
-    anythingElseTI = ui.TextInput(
-        label="Anything else?",
-        style=discord.TextStyle.long,
-        required=False,
-        max_length=1024,
-    )
-
-    async def on_submit(self, interaction: discord.Interaction):
-        respChannel = self.bot.get_channel(968345000100384788)
-        await respChannel.send("")
-
-
-class GSuiteButton(discord.ui.View):
-    def __init__(self, bot):
-        super().__init__(timeout=None)
-        self.value = None
-        self.bot = bot
-
-    @discord.ui.button(
-        label="Get GSuite Account",
-        style=discord.ButtonStyle.blurple,
-        custom_id="persistent_view:gsuite_form",
-        emoji="📝",
-    )
-    async def verify(
-        self,
-        interaction: discord.Interaction,
-        button: discord.ui.Button,
-    ):
-        await interaction.response.send_modal(GSuiteForm(self.bot))
-
-
-class RecruitmentForm(ui.Modal, title="Recruitment Request"):
-    def __init__(self, bot: commands.Bot) -> None:
-        super().__init__(timeout=None)
-        self.bot = bot
-        self.add_item(
-            discord.ui.Select(
-                placeholder="How urgent is your request?",
-                options=[
-                    discord.SelectOption(label="LOW", value="LOW"),
-                    discord.SelectOption(label="MED", value="MED"),
-                    discord.SelectOption(label="HIGH", value="HIGH"),
-                    discord.SelectOption(label="CRIT", value="CRIT"),
-                ],
-            )
-        )
-
-        self.add_item(
-            discord.ui.Select(
-                placeholder="What rank is this position?",
-                options=[
-                    discord.SelectOption(label="Leadership", value="Leadership"),
-                    discord.SelectOption(label="Associate", value="Associate"),
-                ],
-            )
-        )
-
-    position_title = ui.TextInput(
-        label="What is the position title?",
-        style=discord.TextStyle.short,
-        max_length=100,
-    )
-
-    num_people = ui.TextInput(
-        label="How many people are needed?",
-        style=discord.TextStyle.short,
-        max_length=5,
-    )
-
-    async def on_submit(self, interaction: discord.Interaction):
-        mgm_server = self.bot.get_guild(core.common.StaffID.g_staff_resources)
-        ticket_category = discord.utils.get(
-            mgm_server.categories, id=StaffID.cat_recruiting_tickets
-        )
-        member = interaction.guild.get_member(interaction.user.id)
-
-        ticket_channel = await ticket_category.create_text_channel(
-            f"{interaction.user.name}-recruitment-request",
-            topic=f"{interaction.user.name} | {interaction.user.id} recruitment-request",
-            reason=f"Requested by {interaction.user.name} recruitment-request",
-        )
-        query = database.MGMTickets.create(
-            ChannelID=ticket_channel.id,
-            authorID=interaction.user.id,
-            createdAt=datetime.now(),
-        )
-        query.save()
-
-        await ticket_channel.set_permissions(
-            discord.utils.get(mgm_server.roles, name="Human Resources"),
-            read_messages=True,
-            send_messages=True,
-            manage_messages=True,
-        )
-        await ticket_channel.set_permissions(
-            member,
-            read_messages=True,
-            send_messages=True,
-        )
-        control_embed = discord.Embed(
-            title="Control Panel",
-            description="To end this ticket, click the lock button!",
-            color=discord.Colour.gold(),
-        )
-        LCB = discord.ui.View()
-        LCB.add_item(
-            ButtonHandler(
-                style=discord.ButtonStyle.green,
-                url=None,
-                disabled=False,
-                label="Lock",
-                emoji="🔒",
-                custom_id="mgm_ch_lock_menu",
-            )
-        )
-        LCM = await ticket_channel.send(
-            interaction.user.mention, embed=control_embed, view=LCB
-        )
-        await LCM.pin()
-        embed_information = discord.Embed(
-            title="Recruitment Request",
-            description=f"User: {member.mention}\n"
-            f"**Position Title:** {self.position_title.value}\n"
-            f"**Ranking:** {self.children[0].value}\n"
-            f"**People Needed:** {self.num_people.value}\n"
-            f"**Urgency:** {self.children[1].value}",
-            color=discord.Colour.gold(),
-        )
-        embed_information.set_author(
-            name=interaction.user.name,
-            icon_url=interaction.user.avatar.url,
-        )
-        embed_information.set_footer(text=f"ID: {interaction.user.id}")
-        await ticket_channel.send(embed=embed_information)
-        await interaction.response.send(
-            "Your ticket has been created!\nYou can view it here: <#{ticket_channel.id}>"
-        )
-
-
-class RecruitmentButton(discord.ui.View):
-    def __init__(self, bot):
-        super().__init__(timeout=None)
-        self.value = None
-        self.bot = bot
-
-    @discord.ui.button(
-        label="Start Recruitment Request",
-        style=discord.ButtonStyle.blurple,
-        custom_id="persistent_view:recruit_form",
-        emoji="📝",
-    )
-    async def verify(
-        self,
-        interaction: discord.Interaction,
-        button: discord.ui.Button,
-    ):
-        await interaction.response.send_modal(RecruitmentForm(self.bot))
 
 
 class AdminAPI(commands.Cog):
